update readme and fix functions
