+++ docs/source/api.md
@@ -4,6 +4,7 @@
 :caption: How to use MagiAttention
 :maxdepth: 2
 
+install.md
 magi_api.md
 
 ```
