+++ magi_attention/functional/flex_flash_attn.py
@@ -341,6 +341,7 @@ def flex_flash_attn_func(
     """
     An interface similar to flash attention that doesn't require distributed environment, dispatch or undispatch.
     Directly call magi_attn_kernel to get attention output and lse. This is faster when you don't need context parallel.
+
     Args:
         q (torch.Tensor): Query tensor.
         k (torch.Tensor): Key tensor.
@@ -349,11 +350,14 @@ def flex_flash_attn_func(
         k_ranges (torch.Tensor): key ranges in the ref attn mask.
         max_seqlen_q (int): Maximum sequence length of q_ranges.
         max_seqlen_k (int): Maximum sequence length of k_ranges.
-        attn_type_map (torch.Tensor): Attention type map with dtype=torch.int32.
-            0: full attention
-            1: causal attention
-            2: inverse causal attention
-            3: bidirectional causal attention
+        attn_type_map (torch.Tensor): Attention type map with dtype=torch.int32. The values specify
+            the attention type for each token:
+
+                - 0: full attention
+                - 1: causal attention
+                - 2: inverse causal attention
+                - 3: bidirectional causal attention
+
         softmax_scale (float): Softmax scale.
         softcap (float): Softcap.
         deterministic (bool): Whether to use deterministic attention.
@@ -362,83 +366,113 @@ def flex_flash_attn_func(
         disable_fwd_atomic_reduction (bool): Whether to disable forward atomic reduction.
             If you can ensure q_ranges has no overlap, you can set this to True for better performance.
             Overlap in q_ranges is defined as: if any two q_ranges have non-empty intersection, then there is overlap.
-            For example, q_ranges = [[0, 15], [10, 20], [20, 30]] has overlap because [0, 15] and [10, 20] intersect.
-            While q_ranges = [[0, 15], [15, 20], [20, 30]] has no overlap.
+            For example, q_ranges = `[[0, 15], [10, 20], [20, 30]]` has overlap because `[0, 15]` and `[10, 20]` intersect.
+            While q_ranges = `[[0, 15], [15, 20], [20, 30]]` has no overlap.
+
     Returns:
-        out (torch.Tensor): Attention output tensor
-        lse (torch.Tensor): Log-sum-exp values with dtype=torch.float32.
+        tuple[torch.Tensor, torch.Tensor]:
+            - out (torch.Tensor): Attention output tensor
+            - lse (torch.Tensor): Log-sum-exp values with dtype=torch.float32.
+
     Shape:
-        q: (num_tokens_q, num_heads, head_dim)
-        k: (num_tokens_kv, num_heads, head_dim)
-        v: (num_tokens_kv, num_heads, head_dim)
-        q_ranges: (num_ranges, 2)
-        k_ranges: (num_ranges, 2)
-        attn_type_map: (num_ranges, )
-        out: (num_tokens_q, num_heads, head_dim)
-        lse: (num_heads, num_tokens_q)
-    NOTE: attn_type_map explanation:
-        (In addition to the textual explanations provided below, feel free to check out our blog for a visual interpretation:
-        https://sandai-org.github.io/MagiAttention/#flex-flash-attn)
-        1. full attention
-            If seqlen_q = 5 and seqlen_k = 2, the full mask is:
+        - q: (num_tokens_q, num_heads, head_dim)
+        - k: (num_tokens_kv, num_heads, head_dim)
+        - v: (num_tokens_kv, num_heads, head_dim)
+        - q_ranges: (num_ranges, 2)
+        - k_ranges: (num_ranges, 2)
+        - attn_type_map: (num_ranges, )
+        - out: (num_tokens_q, num_heads, head_dim)
+        - lse: (num_heads, num_tokens_q)
+
+    Note:
+        The `attn_type_map` explains the semantics of different attention mask types.
+        In addition to the descriptions below, see our blog for a visual explanation:
+        https://sandai-org.github.io/MagiAttention/blog/#flex-flash-attn
+
+        1. Full attention:
+            - If seqlen_q = 5 and seqlen_k = 2::
+
                 1 1
                 1 1
                 1 1
                 1 1
                 1 1
-            If seqlen_q = 2 and seqlen_k = 5, the full mask is:
+
+            - If seqlen_q = 2 and seqlen_k = 5::
+
                 1 1 1 1 1
                 1 1 1 1 1
-            if seqlen_q = 5 and seqlen_k = 5, the full mask is:
+
+            - If seqlen_q = 5 and seqlen_k = 5::
+
                 1 1 1 1 1
                 1 1 1 1 1
                 1 1 1 1 1
                 1 1 1 1 1
                 1 1 1 1 1
-        2: causal attention (bottom-right aligned)
-            If seqlen_q = 5 and seqlen_k = 2, the causal mask is:
+
+        2. Causal attention (bottom-right aligned):
+            - If seqlen_q = 5 and seqlen_k = 2::
+
                 0 0
                 0 0
                 0 0
                 1 0
                 1 1
-            if seqlen_q = 2 and seqlen_k = 5, the causal mask is:
+
+            - If seqlen_q = 2 and seqlen_k = 5::
+
                 1 1 1 1 0
                 1 1 1 1 1
-            if seqlen_q = 5 and seqlen_k = 5, the causal mask is:
+
+            - If seqlen_q = 5 and seqlen_k = 5::
+
                 1 0 0 0 0
                 1 1 0 0 0
                 1 1 1 0 0
                 1 1 1 1 0
                 1 1 1 1 1
-        3: inverse causal attention (top-left aligned)
-            if seqlen_q = 5 and seqlen_k = 2, the inverse causal mask is:
+
+        3. Inverse causal attention (top-left aligned):
+            - If seqlen_q = 5 and seqlen_k = 2::
+
                 1 1
                 0 1
                 0 0
                 0 0
                 0 0
-            if seqlen_q = 2 and seqlen_k = 5, the inverse causal mask is:
+
+            - If seqlen_q = 2 and seqlen_k = 5::
+
                 1 1 1 1 1
                 0 1 1 1 1
-            if seqlen_q = 5 and seqlen_k = 5, the inverse causal mask is:
+
+            - If seqlen_q = 5 and seqlen_k = 5::
+
                 1 1 1 1 1
                 0 1 1 1 1
                 0 0 1 1 1
                 0 0 0 1 1
                 0 0 0 0 1
-        4. bidirectional causal attention (top-left & bottom-right intersect-aligned)
-            bidirectional causal attention mask is an 'and mask' of causal and inverse causal attention.
-            if seqlen_q = 5 and seqlen_k = 2, the bidirectional causal mask is:
+
+        4. Bidirectional causal attention (intersection of causal and inverse causal):
+            This is the element-wise AND of causal and inverse causal masks.
+
+            - If seqlen_q = 5 and seqlen_k = 2::
+
                 0 0
                 0 0
                 0 0
                 0 0
                 0 0
-            if seqlen_q = 2 and seqlen_k = 5, the bidirectional causal mask is:
+
+            - If seqlen_q = 2 and seqlen_k = 5::
+
                 1 1 1 1 0
                 0 1 1 1 1
-            if seqlen_q = 5 and seqlen_k = 5, the bidirectional causal mask is:
+
+            - If seqlen_q = 5 and seqlen_k = 5::
+
                 1 0 0 0 0
                 0 1 0 0 0
                 0 0 1 0 0
